1) Adjusted master code to mine. 2) Have not finished adDetailsViewController
