--- src/main/java/com/sda/controller/HomeController.java
@@ -1,16 +1,0 @@
-package com.sda.controller;
-
-import com.sda.entity.Customer;
-
-/**
- * @author StanislavR
- */
-public class HomeController extends GeneralController {
-
-    public HomeController(Customer customer) {
-        super(customer);
-    }
-
-    public HomeController() {
-    }
-}
