--- src/main/java/com/sda/controller/SeeAllController.java
@@ -1,16 +1,0 @@
-package com.sda.controller;
-
-import com.sda.entity.Customer;
-
-/**
- * @author StanislavR
- */
-public class SeeAllController extends GeneralController {
-
-    public SeeAllController(Customer customer) {
-        super(customer);
-    }
-
-    public SeeAllController() {
-    }
-}
