+++ src/main/java/com/sda/controller/homeViewController.java
@@ -1,0 +1,30 @@
+package com.sda.controller;
+
+import com.sda.entity.Customer;
+import javafx.event.ActionEvent;
+
+/**
+ * @author StanislavR
+ */
+public class homeViewController extends GeneralController {
+
+    public homeViewController(Customer customer) {
+        super(customer);
+    }
+
+    public homeViewController() {
+    }
+
+    public void logoTextMouseClicked(ActionEvent event) {
+
+    }
+
+    public void myAdsButtonPushed(ActionEvent event) {
+
+    }
+
+    public void allAdsButtonPushed(ActionEvent event) {
+
+    }
+
+}
