+++ src/main/java/controller/Demi_AdDetailsViewController.java
@@ -1,48 +1,48 @@
-package controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class Demi_AdDetailsViewController extends Demi_Models {
-
-    @FXML
-    private Label categoryLabel;
-
-    @FXML
-    private Label priceLabel;
-
-    @FXML
-    private Label startDateLabel;
-
-    @FXML
-    private Label endDateLabel;
-
-    @FXML
-    private Label subjectLabel;
-
-    @FXML
-    private Label adTypeLabel;
-
-    @FXML
-    private Label addressLabel;
-
-    @FXML
-    private Label descriptionLabel;
-
-    @Override
-    public void logoTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @Override
-    public void myAdsButtonPushed(ActionEvent event) {
-
-    }
-
-    @Override
-    public void allAdsButtonPushed(ActionEvent event) {
-
-    }
-
-}
+//package controller;
+//
+//import javafx.event.ActionEvent;
+//import javafx.fxml.FXML;
+//import javafx.scene.control.Label;
+//
+//public class Demi_AdDetailsViewController extends Demi_Models {
+//
+//    @FXML
+//    private Label categoryLabel;
+//
+//    @FXML
+//    private Label priceLabel;
+//
+//    @FXML
+//    private Label startDateLabel;
+//
+//    @FXML
+//    private Label endDateLabel;
+//
+//    @FXML
+//    private Label subjectLabel;
+//
+//    @FXML
+//    private Label adTypeLabel;
+//
+//    @FXML
+//    private Label addressLabel;
+//
+//    @FXML
+//    private Label descriptionLabel;
+//
+//    @Override
+//    public void logoTextMouseClicked(ActionEvent event) {
+//
+//    }
+//
+//    @Override
+//    public void myAdsButtonPushed(ActionEvent event) {
+//
+//    }
+//
+//    @Override
+//    public void allAdsButtonPushed(ActionEvent event) {
+//
+//    }
+//
+//}
