--- src/main/java/controller/Demi_HomeViewController.java
@@ -1,21 +1,0 @@
-package controller;
-
-import javafx.event.ActionEvent;
-
-public class Demi_HomeViewController extends Demi_Models {
-
-    @Override
-    public void logoTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @Override
-    public void myAdsButtonPushed(ActionEvent event) {
-
-    }
-
-    @Override
-    public void allAdsButtonPushed(ActionEvent event) {
-
-    }
-}
