--- src/main/java/controller/Demi_LogInViewController.java
@@ -1,58 +1,0 @@
-package controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
-import javafx.scene.text.Text;
-
-public class Demi_LogInViewController extends Demi_Models {
-
-    @FXML
-    private TextField userNameTextField;
-
-    @FXML
-    private PasswordField passwordPasswordField;
-
-    @FXML
-    private Text signUpNowText;
-
-    @FXML
-    private Button signInButton;
-
-    @FXML
-    void passwordPasswordFieldPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    void userNameTextFieldPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    void signInButtonPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    void signUpNowTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @Override
-    public void logoTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @Override
-    public void myAdsButtonPushed(ActionEvent event) {
-
-    }
-
-    @Override
-    public void allAdsButtonPushed(ActionEvent event) {
-
-    }
-}
