--- src/main/java/controller/Demi_Models.java
@@ -1,34 +1,0 @@
-package controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.text.Text;
-
-public abstract class Demi_Models {
-
-    @FXML
-    protected Text logoText;
-
-    @FXML
-    protected Button myAdsButton;
-
-    @FXML
-    protected Button allAdsButton;
-
-    @FXML
-    public void logoTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @FXML
-    public void myAdsButtonPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    public void allAdsButtonPushed(ActionEvent event) {
-
-    }
-
-}
