--- src/main/java/controller/Demi_SignUpViewController.java
@@ -1,67 +1,0 @@
-package controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
-import javafx.scene.text.Text;
-
-public class Demi_SignUpViewController extends Demi_Models {
-
-    @FXML
-    private TextField userNameTextField;
-
-    @FXML
-    private Text signInText;
-
-    @FXML
-    private PasswordField passwordPasswordField;
-
-    @FXML
-    private TextField fullNameTextField;
-
-    @FXML
-    private Button signUpButton;
-
-    @FXML
-    void passwordPasswordFieldPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    void userNameTextFieldPushed(ActionEvent event) {
-
-    }
-
-    @FXML
-    void signUpButtonPushed(ActionEvent event) {
-
-    }
-
-
-    @FXML
-    void signInTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @FXML
-    void fullNameTextFieldPushed(ActionEvent event) {
-
-    }
-
-    @Override
-    public void logoTextMouseClicked(ActionEvent event) {
-
-    }
-
-    @Override
-    public void myAdsButtonPushed(ActionEvent event) {
-
-    }
-
-    @Override
-    public void allAdsButtonPushed(ActionEvent event) {
-
-    }
-}
