--- src/main/resources/Demi_allAdsView.fxml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.*?>
-<?import javafx.scene.media.*?>
-<?import javafx.scene.input.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.Demi_AllAdsViewController">
-   <children>
-      <HBox layoutX="1077.0" layoutY="14.0" spacing="14.0">
-         <children>
-            <Button fx:id="myAdsButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#myAdsButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="My Ads">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-            <Button fx:id="allAdsButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#allAdsButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="All Ads">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-         </children>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-      </HBox>
-      <VBox layoutX="14.0" layoutY="89.0" spacing="15.0">
-         <children>
-            <VBox spacing="5.0">
-               <children>
-                  <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Category">
-                     <font>
-                        <Font name="System Bold" size="15.0" />
-                     </font>
-                     <VBox.margin>
-                        <Insets />
-                     </VBox.margin>
-                  </Text>
-                  <ComboBox fx:id="categoryComboBox" onAction="#categoryComboBoxPushed" prefHeight="31.0" prefWidth="200.0" />
-               </children>
-               <padding>
-                  <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
-               </padding></VBox>
-            <VBox spacing="5.0">
-               <children>
-                  <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Location">
-                     <font>
-                        <Font name="System Bold" size="15.0" />
-                     </font>
-                  </Text>
-                  <ComboBox fx:id="locationComboBox" onAction="#locationComboBoxPushed" prefHeight="31.0" prefWidth="200.0" />
-               </children>
-               <padding>
-                  <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
-               </padding></VBox>
-            <VBox maxWidth="-Infinity" minWidth="-Infinity" spacing="5.0">
-               <children>
-                  <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Price">
-                     <font>
-                        <Font name="System Bold" size="15.0" />
-                     </font></Text>
-                  <TextField fx:id="priceFromTextField" onAction="#priceFromTextFieldPushed" prefWidth="180.0" promptText="From" />
-                  <TextField fx:id="priceToTextField" onAction="#priceToTextFieldPushed" prefWidth="180.0" promptText="To" />
-               </children>
-               <padding>
-                  <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
-               </padding></VBox>
-            <VBox maxWidth="-Infinity" minWidth="-Infinity" spacing="5.0">
-               <children>
-                  <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Date">
-                     <font>
-                        <Font name="System Bold" size="15.0" />
-                     </font>
-                  </Text>
-                  <DatePicker fx:id="startDatePicker" editable="false" onAction="#startDatePickerPushed" prefWidth="180.0" promptText="Start Date" />
-                  <DatePicker fx:id="endDatePicker" editable="false" onAction="#endDatePickerPushed" prefWidth="180.0" promptText="End Date" />
-               </children>
-               <padding>
-                  <Insets bottom="3.0" left="3.0" right="3.0" top="3.0" />
-               </padding>
-            </VBox>
-         </children>
-      </VBox>
-      <TableView fx:id="mainTableView" layoutX="261.0" layoutY="89.0" prefHeight="682.0" prefWidth="1005.0">
-        <columns>
-          <TableColumn fx:id="subjectColumn" prefWidth="-1.0" text="Subject" />
-          <TableColumn fx:id="categoryColumn" minWidth="0.0" prefWidth="-1.0" text="Category" />
-            <TableColumn fx:id="priceColumn" prefWidth="-1.0" text="Price" />
-            <TableColumn fx:id="locationColumn" prefWidth="-1.0" text="Location" />
-            <TableColumn fx:id="startDateColumn" prefWidth="-1.0" text="Start Date" />
-            <TableColumn fx:id="endDateColumn" prefWidth="-1.0" text="End Date" />
-        </columns>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-      </TableView>
-      <Text fx:id="logoText" boundsType="VISUAL" fill="RED" layoutX="17.0" layoutY="45.0" onMouseClicked="#logoTextMouseClicked" strokeType="OUTSIDE" strokeWidth="0.0" text="Helppy" wrappingWidth="122.3828125">
-         <font>
-            <Font name="Brush Script MT Italic" size="52.0" />
-         </font>
-      </Text>
-      <Separator layoutY="72.0" prefWidth="1280.0" />
-   </children>
-</Pane>
