--- src/main/resources/Demi_homeView.fxml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.Demi_HomeViewController">
-   <children>
-      <HBox layoutX="1077.0" layoutY="14.0" prefHeight="41.0" spacing="14.0">
-         <children>
-            <Button fx:id="myAdsButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#myAdsButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="My Ads">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-            <Button fx:id="allAdsButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#allAdsButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="All Ads">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-         </children>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-      </HBox>
-      <Separator layoutY="72.0" prefWidth="1280.0" />
-      <Text fx:id="logoText" boundsType="VISUAL" fill="RED" layoutX="17.0" layoutY="48.0" onMouseClicked="#logoTextMouseClicked" strokeType="OUTSIDE" strokeWidth="0.0" text="Helppy" wrappingWidth="122.3828125">
-         <font>
-            <Font name="Brush Script MT Italic" size="52.0" />
-         </font>
-      </Text>
-   </children>
-</Pane>
