--- src/main/resources/views/homePage.fxml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.text.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.HomePage">
-   <children>
-      <HBox layoutX="218.0" layoutY="14.0" spacing="14.0">
-         <children>
-            <Button fx:id="seeServicesButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#seeServicesButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="See Services">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-            <Button fx:id="addServicesButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#addServicesButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="Add Services">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-            <Button fx:id="logInButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#logInButtonPushed" style="-fx-background-color: springgreen; -fx-border-color: black;" text="Log-in">
-               <font>
-                  <Font name="System Bold Italic" size="18.0" />
-               </font>
-            </Button>
-         </children>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-      </HBox>
-      <Separator layoutY="72.0" prefWidth="600.0" />
-      <Button fx:id="exitButton" layoutX="205.0" layoutY="309.0" mnemonicParsing="false" onAction="#exitButtonPushed" prefHeight="55.0" prefWidth="190.0" text="EXIT">
-         <font>
-            <Font name="System Bold" size="30.0" />
-         </font>
-         <opaqueInsets>
-            <Insets />
-         </opaqueInsets>
-         <padding>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-         </padding>
-      </Button>
-   </children>
-</Pane>
