+++ src/main/resources/views/logInView.fxml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.*?>
 <?import javafx.scene.paint.*?>
 <?import javafx.geometry.*?>
 <?import javafx.scene.control.*?>
@@ -7,7 +8,7 @@
 <?import java.lang.*?>
 <?import javafx.scene.layout.*?>
 
-<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.Demi_LogInViewController">
+<Pane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sda.controller.logInViewController">
    <children>
       <HBox layoutX="992.0" layoutY="14.0" spacing="14.0">
          <children>
@@ -15,11 +16,17 @@
                <font>
                   <Font name="System Bold" size="17.0" />
                </font>
+               <cursor>
+                  <Cursor fx:constant="HAND" />
+               </cursor>
             </Button>
             <Button fx:id="allAdsButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#allAdsButtonPushed" prefHeight="45.0" prefWidth="130.0" style="-fx-background-color: rgb(0, 204, 102); -fx-background-radius: 35;" text="All Ads" textFill="WHITE">
                <font>
                   <Font name="System Bold" size="17.0" />
                </font>
+               <cursor>
+                  <Cursor fx:constant="HAND" />
+               </cursor>
             </Button>
          </children>
          <opaqueInsets>
@@ -31,6 +38,9 @@
          <font>
             <Font name="Brush Script MT Italic" size="52.0" />
          </font>
+         <cursor>
+            <Cursor fx:constant="HAND" />
+         </cursor>
       </Text>
       <Pane layoutX="441.0" layoutY="150.0" prefHeight="554.0" prefWidth="400.0" style="-fx-effect: dropshadow(gaussian, black, 6, 0.6, 0, 0); -fx-border-color: black; -fx-border-width: 0.09; -fx-border-radius: 20;">
          <opaqueInsets>
@@ -53,13 +63,19 @@
                      <font>
                         <Font name="System Bold" size="17.0" />
                      </font>
+                     <cursor>
+                        <Cursor fx:constant="HAND" />
+                     </cursor>
                   </Button>
-                  <Text blendMode="SRC_ATOP" layoutX="66.0" layoutY="321.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Don't have an account?" />
-                  <Text fx:id="signUpNowText" blendMode="SRC_ATOP" fill="#00cc66" layoutX="92.0" layoutY="352.0" onMouseClicked="#signUpNowTextMouseClicked" strokeType="OUTSIDE" strokeWidth="0.0" text="SIGN UP NOW">
+                  <Text blendMode="SRC_ATOP" layoutX="81.0" layoutY="320.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Don't have an account?" />
+                  <Button fx:id="signUpNowButton" blendMode="SRC_ATOP" layoutX="81.0" layoutY="331.0" mnemonicParsing="false" onAction="#signUpNowButtonPushed" style="-fx-background-color: transparent;" text="SIGN UP NOW" textFill="#00cc66">
                      <font>
                         <Font name="System Bold" size="15.0" />
                      </font>
-                  </Text>
+                     <cursor>
+                        <Cursor fx:constant="HAND" />
+                     </cursor>
+                  </Button>
                </children>
             </Pane>
             <Pane blendMode="SRC_ATOP" prefHeight="103.0" prefWidth="400.0" style="-fx-background-color: rgb(0, 204, 102); -fx-background-radius: 20 20 0 0;">
