--- src/main/resources/views/signInView.fxml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--<?import com.gluonhq.charm.glisten.control.TextField?>-->
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="700.0" style="-fx-background-color: #2E3348;" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sda.controller.LoginViewController">
-   <children>
-      <AnchorPane layoutX="-5.0" layoutY="111.0" prefHeight="290.0" prefWidth="706.0" style="-fx-background-color: fafafa;">
-         <children>
-            <TextField fx:id="loginTextField" layoutX="178.0" layoutY="78.0" prefHeight="26.0" prefWidth="310.0" promptText="Login" style="-fx-background-color: #fafafa; -fx-border-radius: 5;" />
-            <PasswordField fx:id="passwordTextField" layoutX="179.0" layoutY="132.0" prefHeight="26.0" prefWidth="310.0" promptText="Password" style="-fx-background-color: #fafafa; -fx-border-radius: 5;" />
-            <Label layoutX="282.0" layoutY="14.0" text="Sign In" textFill="#8c8989">
-               <font>
-                  <Font name="Britannic Bold" size="33.0" />
-               </font>
-            </Label>
-            <Button fx:id="authSignInButton" layoutX="265.0" layoutY="192.0" mnemonicParsing="false" onAction="#authSignInButtonPushed" prefHeight="42.0" prefWidth="138.0" style="-fx-background-color: #2E3348;" text="ENTER" textFill="WHITE">
-               <font>
-                  <Font name="Britannic Bold" size="16.0" />
-               </font>
-            </Button>
-         </children>
-      </AnchorPane>
-      <Button fx:id="signUpButton" layoutX="577.0" layoutY="37.0" mnemonicParsing="false" onAction="#signUpButtonPushed" prefHeight="32.0" prefWidth="103.0" style="-fx-background-color: #f39c63;" text="Sign Up" textFill="WHITE">
-         <font>
-            <Font name="Britannic Bold" size="16.0" />
-         </font>
-      </Button>
-   </children>
-</AnchorPane>
